requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
aiokafka==0.12.0
minio==7.2.0
requests==2.32.5
defusedxml==0.7.1

# Development and linting tools
ruff==0.13.0
mypy==1.18.1
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
